♻️ `director-v2` settings refactoring (#4993)
