+++ packages/service-library/src/servicelib/utils_secrets.py
@@ -2,6 +2,8 @@ import secrets
 import string
 from typing import Final
 
+from pydantic import StrictInt, validate_arguments
+
 MIN_PASSWORD_LENGTH = 30
 _SAFE_SYMBOLS = "!$%*+,-.:=?@^_~"  # avoid issues with parsing, espapes etc
 _ALPHABET: Final = string.digits + _SAFE_SYMBOLS + string.ascii_letters
@@ -44,3 +46,14 @@ def generate_passcode(number_of_digits: int = MIN_PASSCODE_LENGTH) -> str:
 def are_secrets_equal(got: str, expected: str) -> bool:
     """Constant-time evaluation of 'got == expected'"""
     return secrets.compare_digest(got.encode("utf8"), expected.encode("utf8"))
+
+
+@validate_arguments
+def secure_randint(start: StrictInt, end: StrictInt) -> int:
+    """Generate a random integer between start (inclusive) and end (exclusive)."""
+    if start >= end:
+        msg = f"{start=} must be less than {end=}"
+        raise ValueError(msg)
+
+    diff = end - start
+    return secrets.randbelow(diff) + start
