+++ services/director-v2/src/simcore_service_director_v2/api/dependencies/dynamic_services.py
@@ -8,7 +8,7 @@ from models_library.projects_nodes import NodeID
 from servicelib.logging_utils import log_decorator
 from starlette.datastructures import URL
 
-from ...core.settings import DynamicServicesSettings
+from ...core.dynamic_services_settings import DynamicServicesSettings
 from ...modules.director_v0 import DirectorV0Client
 from ...modules.dynamic_services import ServicesClient
 from ...modules.dynamic_sidecar.scheduler import DynamicSidecarsScheduler
