+++ services/director-v2/src/simcore_service_director_v2/api/routes/dynamic_services.py
@@ -25,6 +25,9 @@ from servicelib.json_serialization import json_dumps
 from servicelib.logging_utils import log_decorator
 from servicelib.rabbitmq import RabbitMQClient
 from servicelib.utils import logged_gather
+from simcore_service_director_v2.core.dynamic_services_settings.scheduler import (
+    DynamicServicesSchedulerSettings,
+)
 from starlette import status
 from starlette.datastructures import URL
 from tenacity import RetryCallState, TryAgain
@@ -35,8 +38,7 @@ from tenacity.wait import wait_fixed
 
 from ...api.dependencies.database import get_repository
 from ...api.dependencies.rabbitmq import get_rabbitmq_client
-from ...core.dynamic_sidecar_settings import DynamicSidecarSettings
-from ...core.settings import DynamicServicesSettings
+from ...core.dynamic_services_settings import DynamicServicesSettings
 from ...modules import projects_networks
 from ...modules.db.repositories.projects import ProjectsRepository
 from ...modules.db.repositories.projects_networks import ProjectsNetworksRepository
@@ -140,7 +142,7 @@ async def create_dynamic_service(
 
     #
     if not await is_sidecar_running(
-        service.node_uuid, dynamic_services_settings.DYNAMIC_SIDECAR
+        service.node_uuid, dynamic_services_settings.DYNAMIC_SCHEDULER.SWARM_STACK_NAME
     ):
         await scheduler.add_service(
             service=service,
@@ -216,8 +218,8 @@ async def stop_dynamic_service(
     # take care of stopping cleaning up all allocated resources:
     # services, containers, volumes and networks.
     # Once the service is no longer being tracked this can return
-    dynamic_sidecar_settings: DynamicSidecarSettings = (
-        dynamic_services_settings.DYNAMIC_SIDECAR
+    dynamic_services_scheduler_settings: DynamicServicesSchedulerSettings = (
+        dynamic_services_settings.DYNAMIC_SCHEDULER
     )
 
     def _log_error(retry_state: RetryCallState):
@@ -230,7 +232,7 @@ async def stop_dynamic_service(
     async for attempt in AsyncRetrying(
         wait=wait_fixed(1.0),
         stop=stop_after_delay(
-            dynamic_sidecar_settings.DYNAMIC_SIDECAR_WAIT_FOR_SERVICE_TO_STOP
+            dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_WAIT_FOR_SERVICE_TO_STOP
         ),
         before_sleep=before_sleep_log(logger=logger, log_level=logging.INFO),
         reraise=False,
@@ -272,12 +274,12 @@ async def service_retrieve_data_on_ports(
             node_uuid, director_v0_client
         )
 
-        dynamic_sidecar_settings: DynamicSidecarSettings = (
-            dynamic_services_settings.DYNAMIC_SIDECAR
+        dynamic_services_scheduler_settings: DynamicServicesSchedulerSettings = (
+            dynamic_services_settings.DYNAMIC_SCHEDULER
         )
         timeout = httpx.Timeout(
-            dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
-            connect=dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_CONNECT_TIMEOUT,
+            dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            connect=dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_CONNECT_TIMEOUT,
         )
 
         # this call waits for the service to download data
