+++ services/director-v2/src/simcore_service_director_v2/core/dynamic_services_settings/__init__.py
@@ -1,0 +1,27 @@
+from pydantic import Field
+from settings_library.base import BaseCustomSettings
+
+from .egress_proxy import EgressProxySettings
+from .proxy import DynamicSidecarProxySettings
+from .scheduler import DynamicServicesSchedulerSettings
+from .sidecar import DynamicSidecarSettings
+
+
+class DynamicServicesSettings(BaseCustomSettings):
+    DIRECTOR_V2_DYNAMIC_SERVICES_ENABLED: bool = Field(
+        default=True, description="Enables/Disables the dynamic_sidecar submodule"
+    )
+
+    DYNAMIC_SIDECAR: DynamicSidecarSettings = Field(auto_default_from_env=True)
+
+    DYNAMIC_SCHEDULER: DynamicServicesSchedulerSettings = Field(
+        auto_default_from_env=True
+    )
+
+    DYNAMIC_SIDECAR_PROXY_SETTINGS: DynamicSidecarProxySettings = Field(
+        auto_default_from_env=True
+    )
+
+    DYNAMIC_SIDECAR_EGRESS_PROXY_SETTINGS: EgressProxySettings = Field(
+        auto_default_from_env=True
+    )
