+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/api_client/_public.py
@@ -25,7 +25,9 @@ from servicelib.fastapi.long_running_tasks.client import (
 from servicelib.logging_utils import log_context, log_decorator
 from servicelib.utils import logged_gather
 
-from ....core.dynamic_sidecar_settings import DynamicSidecarSettings
+from ....core.dynamic_services_settings.scheduler import (
+    DynamicServicesSchedulerSettings,
+)
 from ....models.dynamic_services_scheduler import SchedulerData
 from ....modules.dynamic_sidecar.docker_api import get_or_create_networks_ids
 from ..errors import EntrypointContainerNotFoundError
@@ -59,9 +61,9 @@ class SidecarsClient:
         return self._thin_client.client
 
     @cached_property
-    def _dynamic_sidecar_settings(self) -> DynamicSidecarSettings:
-        settings: DynamicSidecarSettings = (
-            self._app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
+    def _dynamic_services_scheduler_settings(self) -> DynamicServicesSchedulerSettings:
+        settings: DynamicServicesSchedulerSettings = (
+            self._app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SCHEDULER
         )
         return settings
 
@@ -305,7 +307,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_WAIT_FOR_CONTAINERS_TO_START,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_WAIT_FOR_CONTAINERS_TO_START,
             progress_callback,
         )
 
@@ -322,7 +324,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_WAIT_FOR_SERVICE_TO_STOP,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_WAIT_FOR_SERVICE_TO_STOP,
             progress_callback,
         )
 
@@ -335,7 +337,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
             _debug_progress_callback,
         )
 
@@ -352,7 +354,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
             progress_callback,
         )
 
@@ -369,7 +371,7 @@ class SidecarsClient:
         transferred_bytes = await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
             _debug_progress_callback,
         )
         return transferred_bytes or 0
@@ -387,7 +389,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
             _debug_progress_callback,
         )
 
@@ -404,7 +406,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_SAVE_RESTORE_STATE_TIMEOUT,
             progress_callback,
         )
 
@@ -417,7 +419,7 @@ class SidecarsClient:
         await self._await_for_result(
             task_id,
             dynamic_sidecar_endpoint,
-            self._dynamic_sidecar_settings.DYNAMIC_SIDECAR_API_RESTART_CONTAINERS_TIMEOUT,
+            self._dynamic_services_scheduler_settings.DYNAMIC_SIDECAR_API_RESTART_CONTAINERS_TIMEOUT,
             _debug_progress_callback,
         )
