+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_scheduler.py
@@ -38,8 +38,9 @@ from servicelib.background_task import cancel_task
 from servicelib.fastapi.long_running_tasks.client import ProgressCallback
 from servicelib.fastapi.long_running_tasks.server import TaskProgress
 
-from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
-from .....core.settings import DynamicServicesSchedulerSettings
+from .....core.dynamic_services_settings.scheduler import (
+    DynamicServicesSchedulerSettings,
+)
 from .....models.dynamic_services_scheduler import SchedulerData, ServiceName
 from ...api_client import SidecarsClient, get_sidecars_client
 from ...docker_api import update_scheduler_data_label
@@ -191,14 +192,14 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
     async def remove_service_sidecar_proxy_docker_networks_and_volumes(
         self, task_progress: TaskProgress, node_uuid: NodeID
     ) -> None:
-        dynamic_sidecar_settings: DynamicSidecarSettings = (
-            self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
+        dynamic_services_scheduler_settings: DynamicServicesSchedulerSettings = (
+            self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SCHEDULER
         )
         await service_remove_sidecar_proxy_docker_networks_and_volumes(
             task_progress=task_progress,
             app=self.app,
             node_uuid=node_uuid,
-            dynamic_sidecar_settings=dynamic_sidecar_settings,
+            swarm_stack_name=dynamic_services_scheduler_settings.SWARM_STACK_NAME,
         )
 
     async def save_service_state(
@@ -346,17 +347,12 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
                 return
 
             # recreate new observation
-            dynamic_sidecar_settings: DynamicSidecarSettings = (
-                self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
-            )
             dynamic_scheduler: DynamicServicesSchedulerSettings = (
                 self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SCHEDULER
             )
             self._service_observation_task[
                 service_name
-            ] = self.__create_observation_task(
-                dynamic_sidecar_settings, dynamic_scheduler, service_name
-            )
+            ] = self.__create_observation_task(dynamic_scheduler, service_name)
 
         logger.debug("Service '%s' marked for removal from scheduler", service_name)
 
@@ -486,7 +482,6 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
 
     def __create_observation_task(
         self,
-        dynamic_sidecar_settings: DynamicSidecarSettings,
         dynamic_scheduler: DynamicServicesSchedulerSettings,
         service_name: ServiceName,
     ) -> asyncio.Task:
@@ -496,7 +491,6 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
                 scheduler=self,
                 service_name=service_name,
                 scheduler_data=scheduler_data,
-                dynamic_sidecar_settings=dynamic_sidecar_settings,
                 dynamic_scheduler=dynamic_scheduler,
             ),
             name=f"{__name__}.observe_{service_name}",
@@ -512,9 +506,6 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
 
     async def _run_trigger_observation_queue_task(self) -> None:
         """generates events at regular time interval"""
-        dynamic_sidecar_settings: DynamicSidecarSettings = (
-            self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
-        )
         dynamic_scheduler: DynamicServicesSchedulerSettings = (
             self.app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SCHEDULER
         )
@@ -533,9 +524,7 @@ class Scheduler(  # pylint: disable=too-many-instance-attributes
                 logger.info("Create observation task for service %s", service_name)
                 self._service_observation_task[
                     service_name
-                ] = self.__create_observation_task(
-                    dynamic_sidecar_settings, dynamic_scheduler, service_name
-                )
+                ] = self.__create_observation_task(dynamic_scheduler, service_name)
 
         logger.info("Scheduler 'trigger observation queue task' was shut down")
