+++ services/director-v2/tests/integration/02/utils.py
@@ -31,7 +31,10 @@ from simcore_service_director_v2.constants import (
     DYNAMIC_PROXY_SERVICE_PREFIX,
     DYNAMIC_SIDECAR_SERVICE_PREFIX,
 )
-from simcore_service_director_v2.core.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.proxy import (
+    DynamicSidecarProxySettings,
+)
+from simcore_service_director_v2.core.dynamic_services_settings.sidecar import (
     DynamicSidecarSettings,
 )
 from simcore_service_director_v2.models.dynamic_services_scheduler import SchedulerData
@@ -230,15 +233,18 @@ async def patch_dynamic_service_url(app: FastAPI, node_uuid: str) -> str:
 
     # pylint: disable=protected-access
     scheduler: DynamicSidecarsScheduler = app.state.dynamic_sidecar_scheduler
-    service_name = scheduler._scheduler._inverse_search_mapping[
+    service_name = scheduler._scheduler._inverse_search_mapping[  # noqa: SLF001
         NodeID(node_uuid)
-    ]  # noqa: SLF001
-    scheduler_data: SchedulerData = scheduler._scheduler._to_observe[
+    ]
+    scheduler_data: SchedulerData = scheduler._scheduler._to_observe[  # noqa: SLF001
         service_name
-    ]  # noqa: SLF001
+    ]
 
     sidecar_settings: DynamicSidecarSettings = (
         app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
+    )
+    dynamic_sidecar_proxy_settings: DynamicSidecarProxySettings = (
+        app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR_PROXY_SETTINGS
     )
 
     async with _disable_create_user_services(scheduler_data):
@@ -255,7 +261,7 @@ async def patch_dynamic_service_url(app: FastAPI, node_uuid: str) -> str:
 
         proxy_published_port = await _get_service_published_port(
             proxy_service_name,
-            target_port=sidecar_settings.DYNAMIC_SIDECAR_PROXY_SETTINGS.DYNAMIC_SIDECAR_CADDY_ADMIN_API_PORT,
+            target_port=dynamic_sidecar_proxy_settings.DYNAMIC_SIDECAR_CADDY_ADMIN_API_PORT,
         )
         assert proxy_published_port is not None, f"{sidecar_settings.json()=}"
