+++ services/director-v2/tests/unit/with_dbs/test_modules_dynamic_sidecar_docker_service_specs.py
@@ -19,12 +19,14 @@ from models_library.service_settings_labels import (
 )
 from models_library.services import RunID, ServiceKeyVersion
 from models_library.wallets import WalletInfo
-from pydantic import BaseModel
 from pytest import MonkeyPatch
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.utils_envs import setenvs_from_dict
 from servicelib.json_serialization import json_dumps
-from simcore_service_director_v2.core.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.scheduler import (
+    DynamicServicesSchedulerSettings,
+)
+from simcore_service_director_v2.core.dynamic_services_settings.sidecar import (
     DynamicSidecarSettings,
 )
 from simcore_service_director_v2.models.dynamic_services_scheduler import SchedulerData
@@ -80,6 +82,13 @@ def mock_env(
 @pytest.fixture
 def dynamic_sidecar_settings(mock_env: dict[str, str]) -> DynamicSidecarSettings:
     return DynamicSidecarSettings.create_from_envs()
+
+
+@pytest.fixture
+def dynamic_services_scheduler_settings(
+    mock_env: dict[str, str]
+) -> DynamicServicesSchedulerSettings:
+    return DynamicServicesSchedulerSettings.create_from_envs()
 
 
 @pytest.fixture
@@ -377,22 +386,18 @@ def test_get_dynamic_proxy_spec(
     minimal_app: FastAPI,
     scheduler_data: SchedulerData,
     dynamic_sidecar_settings: DynamicSidecarSettings,
+    dynamic_services_scheduler_settings: DynamicServicesSchedulerSettings,
     swarm_network_id: str,
     simcore_service_labels: SimcoreServiceLabels,
     expected_dynamic_sidecar_spec: dict[str, Any],
 ) -> None:
     dynamic_sidecar_spec_accumulated = None
 
-    def _dict(model: BaseModel) -> dict[str, Any]:
-        dict_data = model.dict()
-        proxy_settings: dict[str, Any] = dict_data["DYNAMIC_SIDECAR_PROXY_SETTINGS"]
-        # remove key which always changes
-        del proxy_settings["DYNAMIC_SIDECAR_CADDY_ADMIN_API_PORT"]
-        return dict_data
-
-    assert _dict(dynamic_sidecar_settings) == _dict(
-        minimal_app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
+    assert (
+        dynamic_sidecar_settings
+        == minimal_app.state.settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
     )
+
     expected_dynamic_sidecar_spec_model = AioDockerServiceSpec.parse_obj(
         expected_dynamic_sidecar_spec
     )
@@ -406,6 +411,7 @@ def test_get_dynamic_proxy_spec(
         dynamic_sidecar_spec: AioDockerServiceSpec = get_dynamic_sidecar_spec(
             scheduler_data=scheduler_data,
             dynamic_sidecar_settings=dynamic_sidecar_settings,
+            dynamic_services_scheduler_settings=dynamic_services_scheduler_settings,
             swarm_network_id=swarm_network_id,
             settings=cast(SimcoreServiceSettingsLabel, simcore_service_labels.settings),
             app_settings=minimal_app.state.settings,
@@ -485,6 +491,7 @@ async def test_merge_dynamic_sidecar_specs_with_user_specific_specs(
     minimal_app: FastAPI,
     scheduler_data: SchedulerData,
     dynamic_sidecar_settings: DynamicSidecarSettings,
+    dynamic_services_scheduler_settings: DynamicServicesSchedulerSettings,
     swarm_network_id: str,
     simcore_service_labels: SimcoreServiceLabels,
     expected_dynamic_sidecar_spec: dict[str, Any],
@@ -494,6 +501,7 @@ async def test_merge_dynamic_sidecar_specs_with_user_specific_specs(
     dynamic_sidecar_spec: AioDockerServiceSpec = get_dynamic_sidecar_spec(
         scheduler_data=scheduler_data,
         dynamic_sidecar_settings=dynamic_sidecar_settings,
+        dynamic_services_scheduler_settings=dynamic_services_scheduler_settings,
         swarm_network_id=swarm_network_id,
         settings=cast(SimcoreServiceSettingsLabel, simcore_service_labels.settings),
         app_settings=minimal_app.state.settings,
